mini ai project/requirements.txt
requests
matplotlib
python-dotenv
jupyter
groq